requirement.txt
authy==2.2.3
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
Django==2.2.2
django-silk==3.0.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
geographiclib==1.49
geopy==1.20.0
gprof2dot==2017.9.19
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
moviepy==1.0.0
mysqlclient==1.4.2.post1
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
proglog==0.1.9
pycodestyle==2.5.0
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
python-dateutil==2.8.0
python-slugify==3.0.2
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
tqdm==4.32.2
urllib3==1.25.3
